+++ AnalisisNumerico/AnalisisNumerico.UI/Gauss-Jordan/Frm_ParametrosGaussJordan.cs
@@ -89,7 +89,7 @@ namespace AnalisisNumerico.UI.Gauss_Jordan
             {
                 for (int c = 0; c < columnas; c++)
                 {
-                    parametro.Matriz[i, c] = Convert.ToDecimal(Matriz[i, c].Text);
+                    parametro.Matriz[i, c] = Convert.ToDouble(Matriz[i, c].Text);
                 }
             }
